+++ src/File.php
@@ -18,6 +18,132 @@ namespace Josantonius\File;
  */
 class File {
 
+    /**
+     * Directory separator.
+     *
+     * @since 1.1.3
+     *
+     * @var string
+     */
+    const DS = DIRECTORY_SEPARATOR;
+
+    /**
+     * Check if a file exists in a path or url.
+     *
+     * @since 1.1.3
+     *
+     * @param string $file → path or file url
+     *
+     * @return boolean
+     */
+    public static function exists($file) {
+
+        if (filter_var($file, FILTER_VALIDATE_URL)) {
+            
+            $stream = stream_context_create(['http' => ['method' => 'HEAD']]);
+
+            $content = fopen($file, 'r', null, $stream);
+
+            $headers = stream_get_meta_data($content);
+
+            fclose($fp);
+
+            $status = substr($headers['wrapper_data'][0], 9, 3);
+            
+            return ($status >= 200 && $status < 400);
+        }
+
+        return (file_exists($file) && is_file($file));
+    }
+
+    /**
+     * Delete file.
+     *
+     * @since 1.1.3
+     *
+     * @param string $file → file path
+     *
+     * @return boolean
+     */
+    public static function delete($file) {
+
+        return (self::exists($file) && unlink($file));
+    }
+
+    /**
+     * Create directory.
+     *
+     * @since 1.1.3
+     *
+     * @param string $path → path where to create directory
+     *
+     * @return boolean
+     */
+    public static function createDir($path) {
+
+        return (!is_dir($path) && mkdir($path, 0777, true));
+    }
+
+    /**
+     * Delete empty directory.
+     *
+     * @since 1.1.3
+     *
+     * @param string $path → path to delete
+     *
+     * @return boolean
+     */
+    public static function deleteEmptyDir($path) {
+
+        return (is_dir($path) && rmdir($path));
+    }
+
+    /**
+     * Delete directory recursively.
+     *
+     * @since 1.1.3
+     *
+     * @param string $_path → path to delete
+     *
+     * @return boolean
+     */
+    public static function deleteDirRecursively($_path) {
+
+        if (!$path = self::getFilesFromDir($_path)) { return false; }
+
+        foreach($path as $file) {
+
+            if ($file->isFile()) {
+
+                if (!self::delete($file->getRealPath())) { return false; }
+            
+            } else if (!$file->isDot() && $file->isDir()) {
+                
+                self::deleteDirRecursively($file->getRealPath());
+
+                self::deleteEmptyDir($file->getRealPath());
+            }
+        }
+
+        return self::deleteEmptyDir($_path);
+    }
+
+    /**
+     * Get files from directory.
+     *
+     * @since 1.1.3
+     *
+     * @param string $path → path where get files
+     *
+     * @return object|false → 
+     */
+    public static function getFilesFromDir($path) {
+
+        if (!is_dir($path)) { return false; }
+
+        return new \DirectoryIterator(rtrim($path, self::DS) . self::DS);
+    }
+    
    /**
      * Search for part of a string in a file.
      *
@@ -26,7 +152,7 @@ class File {
      * @param string $search   → word or term to search
      * @param string $pathfile → full path to the file
      *
-     * @return bool
+     * @return boolean
      */
     public static function searchString($search, $pathfile) {